@@ -29,41 +29,38 @@
   "license": "ISC",
   "dependencies": {
     "@openapitools/openapi-generator-cli": "2.1.26",
-    "chokidar": "^3.0.2",
-    "copyfiles": "^2.1.0",
-    "cross-spawn": "^6.0.5",
-    "css-loader": "^5.1.3",
-    "css-minimizer-webpack-plugin": "^1.3.0",
+    "chokidar": "^3.5.3",
+    "copyfiles": "^2.4.1",
+    "cross-spawn": "^7.0.3",
+    "css-loader": "^6.6.0",
+    "css-minimizer-webpack-plugin": "^3.4.1",
     "eol-converter-cli": "^1.0.8",
     "file-loader": "^6.2.0",
-    "generic-pool": "latest",
-    "less": "^3.8.1",
+    "generic-pool": "^3.8.2",
+    "less": "^3.13.1",
     "less-loader": "^7.3.0",
-    "meow": "^5.0.0",
-    "mini-css-extract-plugin": "^1.3.9",
-    "raw-loader": "^1.0.0",
-    "rimraf": "^2.6.2",
-    "sass": "^1.43.4",
-    "sass-loader": "^10.1.1",
-    "source-map-loader": "^0.2.4",
-    "uglifyjs-webpack-plugin": "^2.2.0",
-    "webpack": "^4.37.0",
-    "xml2js": "^0.4.19"
+    "meow": "^9.0.0",
+    "mini-css-extract-plugin": "^1.6.2",
+    "rimraf": "^3.0.2",
+    "sass": "^1.49.7",
+    "sass-loader": "^12.4.0",
+    "source-map-loader": "^3.0.1",
+    "webpack": "^5.68.0",
+    "xml2js": "^0.4.23"
   },
   "devDependencies": {
-    "@types/copyfiles": "^2.1.0",
-    "@types/cross-spawn": "^6.0.0",
-    "@types/generic-pool": "^3.1.6",
+    "@types/copyfiles": "^2.4.1",
+    "@types/cross-spawn": "^6.0.2",
+    "@types/generic-pool": "^3.1.10",
     "@types/jest": "^27.0.3",
-    "@types/less": "^3.0.0",
-    "@types/meow": "^5.0.0",
-    "@types/minimist": "^1.2.0",
-    "@types/node": "^10.12.10",
-    "@types/rimraf": "^2.0.2",
-    "@types/vinyl": "^2.0.2",
-    "@types/webpack": "^4.4.19",
-    "jest": "^27.4.5",
-    "ts-jest": "^27.1.2",
+    "@types/less": "^3.0.3",
+    "@types/minimist": "^1.2.2",
+    "@types/node": "^14.18.11",
+    "@types/rimraf": "^3.0.2",
+    "@types/vinyl": "^2.0.6",
+    "@types/webpack": "^5.28.0",
+    "jest": "^27.5.1",
+    "ts-jest": "^27.1.3",
     "typescript": "^3.9.10",
     "tmp": "^0.2.1"
   }
